codicons: Split list into Library and Derived
